+++ ttnn/cpp/ttnn/tensor/page_config.hpp
@@ -1,0 +1,58 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <array>
+
+namespace tt::tt_metal {
+
+class RowMajorPageConfig {
+public:
+    Alignment create_default_alignment(DataType dataType) const;
+    void validate_alignment(const Alignment& alignment, DataType dataType) const;
+
+    Size get_page_shape(const Size& physical_size, const MemoryConfig& memoryConfig) const;
+    size_t get_page_size_bytes(const Size& page_size, DataType dataType) const;
+};
+
+class TilePageConfig {
+public:
+    TilePageConfig(const Tile& tile = Tile());
+
+    Alignment create_default_alignment(DataType dataType) const;
+    void validate_alignment(const Alignment& alignment, DataType dataType) const;
+
+    Size get_page_shape(const Size& physical_size, const MemoryConfig& memoryConfig) const;
+    size_t get_page_size_bytes(const Size& page_size, DataType dataType) const;
+
+    const Tile& get_tile() const;
+
+private:
+    Tile mTile;
+};
+
+class PageConfig {
+public:
+    using Config = std::variant<RowMajorPageConfig, TilePageConfig>;
+
+    PageConfig(const Config& config);
+    PageConfig(Layout layout);
+    PageConfig(Layout layout, const std::optional<Tile>& tile);
+
+    Alignment create_default_alignment(DataType dataType) const;
+    void validate_alignment(const Alignment& alignment, DataType dataType) const;
+
+    Size get_page_shape(const Size& physical_size, const MemoryConfig& memoryConfig) const;
+    size_t get_page_size_bytes(const Size& page_size, DataType dataType) const;
+
+    std::optional<Tile> get_tile() const;
+
+    bool is_row_major() const;
+
+private:
+    Config mConfig;
+};
+
+} // namespace tt::tt_metal
