+++ ttnn/cpp/ttnn/tensor/types.cpp
@@ -338,49 +338,6 @@ MemoryConfig load_memory_config(const std::string& file_name) {
 
 }  // namespace tt
 
-namespace ttnn {
-
-namespace {
-int32_t normalized_index(int32_t index, size_t container_size) {
-    int32_t size = static_cast<int32_t>(container_size);
-
-    if (index < 0) {
-        index += size;
-    }
-
-    if (index < 0 || index >= size) {
-        throw std::out_of_range("SimpleShape index out of range.");
-    }
-
-    return index;
-}
-}
-
-bool SimpleShape::operator==(const SimpleShape &other) const {
-    return this->value == other.value;
-}
-
-bool SimpleShape::operator==(const std::vector<uint32_t> &other) const {
-    return this->value == other;
-}
-
-uint32_t SimpleShape::operator[](int32_t index) const {
-    auto norm_index = normalized_index(index, value.size());
-    return value[norm_index];
-}
-
-uint32_t& SimpleShape::operator[](int32_t index) {
-    auto norm_index = normalized_index(index, value.size());
-    return value[norm_index];
-}
-
-uint64_t SimpleShape::volume() const {
-    return std::accumulate(this->value.begin(), this->value.end(),
-                           uint64_t{1}, std::multiplies<uint64_t>());
-}
-
-} // namespace ttnn
-
 namespace ttnn::types {
 
 uint32_t Shape::operator[](std::int64_t index) const { return this->value.without_padding()[index]; }
